Move the cli input / output method tests to their own test module.
